improve auth

+++ Acadenote.Server/Controllers/NotesController.cs
@@ -1,5 +1,6 @@
 ﻿using Acadenode.Core.Models;
 using Acadenode.Core.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -7,6 +8,7 @@ using System.Threading.Tasks;
 namespace Acadenote.Server.Controllers
 {
     [ApiController]
+    [Authorize]
     [Route("api/[controller]")]
     public class NotesController : ControllerBase
     {
@@ -19,6 +21,7 @@ namespace Acadenote.Server.Controllers
             _logger = logger;
         }
 
+        [AllowAnonymous]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Note>>> GetAllNotesAsync()
         {
@@ -34,6 +37,7 @@ namespace Acadenote.Server.Controllers
             }
         }
 
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public async Task<ActionResult<Note>> GetNoteByIdAsync(string id)
         {
@@ -54,16 +58,22 @@ namespace Acadenote.Server.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> AddNoteAsync(Note note)
+        public async Task<ActionResult> AddNoteAsync([FromBody] Note note)
         {
+            if(!Utils.TryGetJwtToken(Request , out string token) && !Utils.GetRoleFromJwtToken(token).HasAnyRole(Role.Admin, Role.Writer))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var response = await _noteRepository.AddNoteAsync(note);
                 if (!response.Success)
                 {
                     _logger.LogError("Failed to add note: {Error}", response.Message);
-                    return StatusCode(500, $"Failed to add note: {response.Message}");
+                    return BadRequest(response.Message);
                 }
+
                 return CreatedAtAction(nameof(GetNoteByIdAsync), new { id = note.Id }, note);
             }
             catch (Exception ex)
@@ -76,6 +86,11 @@ namespace Acadenote.Server.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateNoteAsync(string id, Note note)
         {
+
+            if (!Utils.TryGetJwtToken(Request, out string token) && !Utils.GetRoleFromJwtToken(token).HasAnyRole(Role.Admin, Role.Writer))
+            {
+                return Unauthorized();
+            }
             try
             {
                 if (id != note.Id)
@@ -102,6 +117,11 @@ namespace Acadenote.Server.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteNoteAsync(string id)
         {
+
+            if (!Utils.TryGetJwtToken(Request, out string token) && !Utils.GetRoleFromJwtToken(token).HasAnyRole(Role.Admin, Role.Writer))
+            {
+                return Unauthorized();
+            }
             try
             {
                 var response = await _noteRepository.DeleteNoteAsync(id);
